+++ frontend/src/components/Skeleton/DrinksDetailDescriptionSkeleton.tsx
@@ -1,5 +1,5 @@
-import FlexBox from 'src/components/FlexBox/FlexBox';
-import Skeleton from './Skeleton';
+import FlexBox from 'src/components/@shared/FlexBox/FlexBox';
+import Skeleton from '../@shared/Skeleton/Skeleton';
 
 const DrinksDetailDescriptionSkeleton = () => {
   return (
